backend/services/text-extraction-service/pyproject.toml
[project]
name = "text-extraction-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Jonathan Hop",email = "yithril@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "pypdf (>=4.0.0,<5.0.0)",
    "reportlab (>=4.0.0,<5.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "striprtf (>=0.0.29,<0.0.30)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "azure-storage-blob (>=12.26.0,<13.0.0)",
    "common-lib @ file:///C:/Coding/LDA-Project/shared/common-lib",
    "shared-auth @ file:///C:/Coding/LDA-Project/shared/auth",
    "csvkit (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.17.0"


[tool.poetry.dependencies]
common-lib = {develop = true}
shared-auth = {develop = true}
